Scroll to hash on page mount
