+++ src/router/scroll_to_hash.js
@@ -1,0 +1,13 @@
+function scroll_to_hash() {
+  if(window.location.hash) {
+    /* URL with hash exists */
+    let id = window.location.hash.substring(1);
+    let element = document.getElementById(id);
+    if(element) {
+      /* Element with hash exists */
+      element.scrollIntoView();
+    }
+  }
+}
+
+export default scroll_to_hash;
